requirements.txt
fastapi>=0.115.0
uvicorn>=0.27.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.5
pydantic>=2.0.0
pydantic-settings>=2.0.0
alembic>=1.12.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
email-validator>=2.0.0
httpx>=0.24.0

requests>=2.31.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
psycopg2-binary>=2.9.5

